scr: make MPI_BUF_SIZE type consistent with components
used as buffer size to MPI and MPI allows at most intfor the buffer
count

+++ src/scr_globals.c
@@ -108,7 +108,7 @@ int scr_set_size      = SCR_SET_SIZE;     /* specify number of tasks in redundan
 int scr_num_encs      = SCR_NUM_ENCS;     /* specify number of encoding blocks in redundancy set */
 int scr_cache_bypass  = SCR_CACHE_BYPASS; /* default bypass, whether to directly read/write parallel file system */
 
-size_t scr_mpi_buf_size  = SCR_MPI_BUF_SIZE;  /* set MPI buffer size to chunk file transfer */
+int scr_mpi_buf_size  = SCR_MPI_BUF_SIZE;     /* set MPI buffer size to chunk file transfer */
 size_t scr_file_buf_size = SCR_FILE_BUF_SIZE; /* set buffer size to chunk file copies to/from parallel file system */
 int scr_copy_metadata    = SCR_COPY_METADATA; /* whether file metadata should also be copied */
